refactor(apps/api): reorganize library-albums DTOs into request/response subdirectories
- Move request DTOs to dto/request/ subdirectory with library-album prefix
- Move response DTOs to dto/response/ subdirectory with library-album prefix
- Update all import paths in commands and controller
- Improve DTO organization and naming consistency
