+++ apps/api/src/features/library-albums/dto/request/create-library-album.request.dto.ts
