+++ apps/api/src/features/library-albums/dto/request/upload-library-album-cover.request.dto.ts
