+++ apps/api/src/features/library-albums/dto/response/create-library-album.response.dto.ts
