+++ apps/api/src/features/library-albums/dto/response/get-library-album.response.dto.ts
