+++ apps/api/src/features/library-albums/dto/response/update-library-album.response.dto.ts
