+++ apps/api/src/features/library-albums/dto/response/upload-library-album-cover.response.dto.ts
